# v2: add Hex helpers and dont quote values with newlines