To ensure that calls like spew.DumpS() can be used to clearly visualise
the contents of an object, we prevent adding quotes to such attribute
objects.