@ellemouton 2024-12-10
Before and after benchmarks: Slight impact on S logs only since we now need to do the extra check on each value:

```
BEFORE

BenchmarkLogger/Skipped_Simple_`S`_Log-10         	476279935	         2.491 ns/op
BenchmarkLogger/Complex_`S`_Log-10                	  307382	      3711 ns/op

AFTER 

BenchmarkLogger/Simple_`S`_Log-10                 	 4641754	       264.2 ns/op
BenchmarkLogger/Complex_`S`_Log-10                	  302826	      3821 ns/op
```