@ellemouton 2025-01-02
cc @Roasbeef  (we need this for https://github.com/lightningnetwork/lnd/pull/9342)